Add Compra de cryptos

+++ src/main/java/br/com/fiap/carteiracryptosspring/dto/CryptoClienteDTO.java
@@ -1,10 +1,12 @@
 package br.com.fiap.carteiracryptosspring.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class CryptoClienteDTO {
    
    private String codigo;
-   private Double quantidade;
+   private BigDecimal quantidade;
 }
